==> data_aug/2_korquad/02.korquad_analysis.ipynb <==
from datasets import load_dataset

dataset = load_dataset("KorQuAD/squad_kor_v1")

dataset

import pandas as pd

df_train = pd.DataFrame(dataset['train'])

df_valid = pd.DataFrame(dataset['validation'])

# 1차 필터링 (LLM으로 context만 판별)

## train dataset에서 유효 id를 가진 row만 추출
train dataset: 60407 -> 9041

df = pd.read_csv('korquad_train_cls.csv')
df = df.drop(columns=['Unnamed: 0'])
df_ok = df[df['cls']=='유효']
df_ok_ids = df_ok['id'].to_list()

# 리스트에서 앞의 두 구간 추출
id_prefix_set = {f"{id.split('-')[0]}-{id.split('-')[1]}" for id in df_ok_ids}

# DataFrame에서 동일한 패턴 추출
df_train['prefix'] = df_train['id'].apply(lambda x: f"{x.split('-')[0]}-{x.split('-')[1]}")

# 필터링
filtered_df = df_train[df_train['prefix'].isin(id_prefix_set)].drop(columns=['prefix'])

filtered_df.to_csv('korquad_train_remain.csv')

## valid dataset에서 유효 id를 가진 row만 추출
valid dataset: 5774 -> 898

df = pd.read_csv('korquad_valid_cls.csv')
df = df.drop(columns=['Unnamed: 0'])
df_ok = df[df['cls']=='유효']

df_ok_ids = df_ok['id'].to_list()

# 리스트에서 앞의 두 구간 추출
id_prefix_set = {f"{id.split('-')[0]}-{id.split('-')[1]}" for id in df_ok_ids}

# DataFrame에서 동일한 패턴 추출
df_valid['prefix'] = df_valid['id'].apply(lambda x: f"{x.split('-')[0]}-{x.split('-')[1]}")

# 필터링
filtered_df = df_valid[df_valid['prefix'].isin(id_prefix_set)].drop(columns=['prefix'])

filtered_df.to_csv('korquad_valid_remain.csv')

## concat

df_train_reamain = pd.read_csv('korquad_train_remain.csv')
df_valid_reamain = pd.read_csv('korquad_valid_remain.csv')

df_reamin = pd.concat([df_train_reamain, df_valid_reamain])
df_reamin = df_reamin.drop(columns=['Unnamed: 0'])

df_reamin.to_csv('../resources/aug/korquad/korquad_cls.csv')
